requirements.txt
Django==2.2
pymysql
mysqlclient
django-mysql
djangorestframework
Pillow
